Migrate javascript to webpack pipeline
This pulls in the webpack asset pipline, and begins to migrate away from
sprockets. It includes sweeping changes due to some out of order
thinking as I was handling the migration, but everything is horribly
tied together and I can't really pick it apart.

Biggest change noticable to developer is inclusion of yarn in the build
process, which adds an additional package definition file (package.json)
and lockfile (yarn.lock)

Docker configuration gets more complicated. The node modules folder
needs to be explicitly *not* bind-mounted or pulled in from the
developer's machine to the rails container, which is handled both by
an anonymous volume mount in the docker-compose.yml file and the
addition of a .dockerignore